fix(eco-store): #86c99rjxt BUG-005 add auth guard on /perfil with returnUrl flow
Wire shared ecoStoreAuthGuard on /perfil (canActivateChild) and reuse it
on /comandes, replacing the per-domain not-logged-orders guard. Login
facade reads a sanitized returnUrl query param so deep-links survive
the login round-trip. Cart not-logged-shipping confirm dialog also
forwards returnUrl on its "Login" branch.
