+++ libs/eco-store/cart/feature/src/lib/guards/not-logged-shipping.guard.ts
@@ -2,11 +2,14 @@ import { map, Observable, of, switchMap, take } from 'rxjs';
 
 import { inject } from '@angular/core';
 import { toObservable } from '@angular/core/rxjs-interop';
-import { CanActivateFn, Router, UrlTree } from '@angular/router';
+import { CanActivateFn, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
 import { pocketBaseUserProfileStore } from '@plastik/auth/pocketbase/data-access';
 import { SharedConfirmDialogService } from '@plastik/shared/confirm';
 
-export const ecoStoreNotLoggedShippingGuard: CanActivateFn = (): Observable<boolean | UrlTree> => {
+export const ecoStoreNotLoggedShippingGuard: CanActivateFn = (
+  _route,
+  state: RouterStateSnapshot
+): Observable<boolean | UrlTree> => {
   const profileStore = inject(pocketBaseUserProfileStore);
   const confirmService = inject(SharedConfirmDialogService);
   const router = inject(Router);
@@ -26,7 +29,9 @@ export const ecoStoreNotLoggedShippingGuard: CanActivateFn = (): Observable<bool
             take(1),
             map(confirmed => {
               if (confirmed) {
-                return router.createUrlTree(['/accedir']);
+                return router.createUrlTree(['/accedir'], {
+                  queryParams: { returnUrl: state.url },
+                });
               }
               return router.createUrlTree(['/cistella/resum'], { relativeTo: null });
             })
