+++ libs/eco-store/core/router/src/index.ts
@@ -1,2 +1,3 @@
+export * from './lib/eco-store-auth.guard';
 export * from './lib/eco-store-category-route-title.service';
 export * from './lib/eco-store-prefix-title.service';
