+++ libs/eco-store/orders/data-access/src/index.ts
@@ -1,2 +1,1 @@
 export * from './eco-store-orders.store';
-export * from './not-logged-orders.guard';
