+++ libs/eco-store/orders/feature/list/src/lib/eco-store-orders-list.routes.ts
@@ -1,5 +1,4 @@
 import { Route } from '@angular/router';
-import { ecoStoreNotLoggedOrdersGuard } from '@plastik/eco-store/orders/data-access';
 import { provideInputSearchFormly } from '@plastik/shared/form/input-search';
 import { provideEcoStoreOrdersFormly } from './eco-store-orders-formly.providers';
 import EcoStoreOrdersListComponent from './eco-store-orders-list.component';
@@ -11,7 +10,6 @@ export const ecoStoreOrdersListRoutes: Route[] = [
     title: 'orders.list.title',
     component: EcoStoreOrdersListComponent,
     providers: [provideEcoStoreOrdersFormly(), provideInputSearchFormly()],
-    canActivate: [ecoStoreNotLoggedOrdersGuard],
     resolve: { ready: ecoStoreOrdersListResolver },
     runGuardsAndResolvers: 'paramsOrQueryParamsChange',
   },
